android: Remove discontiguous planes check from camera buffer
The way discontiguous planes are checked in generic_camera_buffer
ignores that they can be placed in the same dmabuf. This is the
case for AM62x boards, where multiplanar formats are stored
continuously in the memory and referenced through only one
File Descriptor.

Remove the check from generic_camera_buffer.cpp to allow the
creation of CameraBuffers for AM62x boards. The verification
for discontiguous planes are done in v4l2_videodevice.cpp.

Signed-off-by: Vitor Sato Eschholz <vsatoes@baylibre.com>
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>